Create filterclub-logo-sticker-with-background.svg
Add background to logo for screen use. The borders become more pronounced and the color compliments the main brand color and its tints. This is intended to be used for social media accounts. It crops neatly for Instagram / Facebook, leaving just the right amount of white space around the main cone shape.